requirements.txt
pdfminer.six
python-docx
pytesseract
pdf2image
Pillow
pdfplumber